Check the chips on boards nobody built for them — and fix the matcher
Both verifiers stage a chip onto a board THIS REPO BUILDS, choosing filler
that provably cannot take part. That staging is the right way to ask whether
a chip is true, and it was also the whole of the remaining doubt: a live
board's don't-care cells hold real colours, and real colours can join a
match. Nothing measured that gap; "verified" quietly meant "verified against
our own imagination".

  capture_boards.js     3,320 distinct settled boards the shipped bot
                        actually played, level 10, 8 deterministic seeds
  realboards.json       that fixture, one char per cell (243KB, not 800)
  chips.realboard.test.js  find chips that genuinely occur on those boards,
                        fire their own swaps on a live Stack settled to
                        stillness, compare against the claim

MEASURED, whole library, boards nobody built for it:

  145  distinct chips ever located — 2.6% of 5,672. Not a fault in the
       chips: those boards simply did not contain the other shapes. It is a
       statement about FREQUENCY, and it matters for whether a chip-based
       feature would ever fire, not for whether chips are true.
   94% of located chips cleared something
   88% hit the claimed chain depth exactly
   74% hit chain AND panel count exactly

So the staging is substantially honest, which is what needed establishing.

AND THE MEASUREMENT IMMEDIATELY FOUND A BUG IN THE THING DOING THE
MEASURING. The first run said 3,538 chips matched a SINGLE board and 92% of
them cleared nothing. The number was the tell. chipmatch.js kept its
slot/colour bookkeeping in Int8Array while stamping it with an ever-rising
generation counter, so from the 128th call on the stamp comparison could
never be true and the two constraints that give a template its teeth — cells
in one slot share a colour, different slots do not — stopped being enforced
at all. The matcher had degenerated into "are these cells non-empty".

It survived because the test nearest it could not see it: hookup asks the
matcher to find a chip on a board built FROM that chip, where there is one
placement and one colour per slot and nothing for a constraint to reject.
So chipmatch.test.js now pins all four constraints directly, both
directions, running each 200 times so a byte-sized counter is inside the
window rather than just outside it.

Both new gates proved able to fail: restoring the Int8Array fails the
real-board gate on two of its four checks.

Gates added: chips hold up on real boards, and the matcher enforces every
constraint. gate_all exits 0.

+++ games/the-game/ai/eval/chipmatch.js
@@ -132,9 +132,26 @@
     // Slot -> colour for one placement. Reused across placements rather than
     // reallocated: this runs a few million times a second and a fresh object
     // per placement was most of the cost.
-    var slotOf = new Int8Array(16);
-    var slotSeen = new Int8Array(16);
-    var colourTaken = new Int8Array(16);
+    // INT32, NOT INT8, AND THIS WAS A REAL BUG. `stamp` is an ever-rising
+    // generation counter and these arrays hold the stamp a slot was last
+    // written at. In an Int8Array the stored value truncates at 127, so from
+    // the 128th call onward `slotSeen[s] === stamp` was comparing a truncated
+    // byte against a full number and was essentially never true — which meant
+    // the two constraints that give a template its teeth stopped being
+    // enforced at all: cells sharing a colour slot no longer had to be the
+    // same colour, and two different slots no longer had to be different
+    // colours. The matcher degenerated into "are these cells non-empty".
+    //
+    // It matched 3,538 chips on a single real board and 92% of them cleared
+    // NOTHING when fired. The tell was the number, not the code: a library
+    // where nearly every shape is present on every board is not a library.
+    // chipmatch.test.js pins the two constraints directly, and the real-board
+    // measurement is what surfaced it — matching a chip against a board built
+    // from that same chip passes either way, because there is only one
+    // placement and one colour per slot to get wrong.
+    var slotOf = new Int32Array(16);
+    var slotSeen = new Int32Array(16);
+    var colourTaken = new Int32Array(16);
     var stamp = 0;
 
     function matchAt(grid, tmpl, R, C) {
